+++ src/Core/Node.cpp
@@ -214,7 +214,7 @@ void Node::P2PClientBytecoin::on_msg_notify_new_block(NOTIFY_NEW_BLOCK::request 
 	switch (action) {
 	case BroadcastAction::BAN:
 		disconnect("NOTIFY_NEW_BLOCK add_block BAN");
-		return;
+        return;
 	case BroadcastAction::BROADCAST_ALL: {
 		req.hop += 1;
 		BinaryArray raw_msg = LevinProtocol::send_message(NOTIFY_NEW_BLOCK::ID, LevinProtocol::encode(req), false);